11  - Capstone project/requirements.txt
streamlit==1.29.0
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.6.1
langchain==0.1.5
langchain-openai==0.0.2.post1
langchain-community==0.0.20